requirements.txt
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
youtube-transcript-api==0.6.2
requests==2.31.0
openai-whisper 
ffmpeg-python
yt-dlp